fix(test): broken test, due to refactoring
