+++ lua/neotest-golang/results_dir.lua
@@ -5,22 +5,6 @@ local json = require("neotest-golang.json")
 
 local M = {}
 
--- Find the common path of two folderpaths.
-function M.find_common_path(path1, path2)
-  local common = {}
-  local path1_parts = vim.split(path1, "/")
-  local path2_parts = vim.split(path2, "/")
-  for i = #path1_parts, 1, -1 do
-    if path1_parts[i] == path2_parts[#path2_parts] then
-      table.insert(common, 1, path1_parts[i])
-      table.remove(path2_parts)
-    else
-      break
-    end
-  end
-  return table.concat(common, "/")
-end
-
 ---@param spec neotest.RunSpec
 ---@param result neotest.StrategyResult
 ---@param tree neotest.Tree
